requirements.txt
requests
beautifulsoup4
streamlit
google-generativeai
python-dotenv
sqlalchemy
